Move TypeInfo to a separate header + objtreelite_common
