+++ include/treelite/base.h
@@ -12,6 +12,7 @@
 #include <string>
 #include <unordered_map>
 #include <stdexcept>
+#include "./typeinfo.h"
 
 namespace treelite {
 
@@ -31,21 +32,8 @@ enum class Operator : int8_t {
   kGE,  /*!< operator >= */
 };
 
-/*! \brief Types used by thresholds and leaf outputs */
-enum class TypeInfo : uint8_t {
-  kInvalid = 0,
-  kUInt32 = 1,
-  kFloat32 = 2,
-  kFloat64 = 3
-};
-static_assert(std::is_same<std::underlying_type<TypeInfo>::type, uint8_t>::value,
-              "TypeInfo must use uint8_t as underlying type");
-
 /*! \brief conversion table from string to Operator, defined in tables.cc */
 extern const std::unordered_map<std::string, Operator> optable;
-
-/*! \brief conversion table from string to TypeInfo, defined in tables.cc */
-extern const std::unordered_map<std::string, TypeInfo> typeinfo_table;
 
 /*!
  * \brief get string representation of comparison operator
@@ -60,46 +48,6 @@ inline std::string OpName(Operator op) {
     case Operator::kGT: return ">";
     case Operator::kGE: return ">=";
     default: return "";
-  }
-}
-
-/*!
- * \brief Get string representation of type info
- * \param info a type info
- * \return string representation
- */
-inline std::string TypeInfoToString(treelite::TypeInfo type) {
-  switch (type) {
-  case treelite::TypeInfo::kInvalid:
-    return "invalid";
-  case treelite::TypeInfo::kUInt32:
-    return "uint32";
-  case treelite::TypeInfo::kFloat32:
-    return "float32";
-  case treelite::TypeInfo::kFloat64:
-    return "float64";
-  default:
-    throw std::runtime_error("Unrecognized type");
-    return "";
-  }
-}
-
-/*!
- * \brief Convert a template type into a type info
- * \tparam template type to be converted
- * \return TypeInfo corresponding to the template type arg
- */
-template <typename T>
-inline TypeInfo InferTypeInfoOf() {
-  if (std::is_same<T, uint32_t>::value) {
-    return TypeInfo::kUInt32;
-  } else if (std::is_same<T, float>::value) {
-    return TypeInfo::kFloat32;
-  } else if (std::is_same<T, double>::value) {
-    return TypeInfo::kFloat64;
-  } else {
-    throw std::runtime_error(std::string("Unrecognized Value type") + typeid(T).name());
-    return TypeInfo::kInvalid;
   }
 }
