+++ src/optable.cc
@@ -1,10 +1,12 @@
 /*!
  * Copyright (c) 2017-2020 by Contributors
- * \file tables.cc
+ * \file optable.cc
  * \author Hyunsu Cho
- * \brief Conversion tables to obtain Operator and TypeInfo from strings
+ * \brief Conversion tables to obtain Operator from string
  */
 
+#include <string>
+#include <unordered_map>
 #include <treelite/base.h>
 
 namespace treelite {
@@ -15,12 +17,6 @@ const std::unordered_map<std::string, Operator> optable{
   {"<=", Operator::kLE},
   {">",  Operator::kGT},
   {">=", Operator::kGE}
-};
-
-const std::unordered_map<std::string, TypeInfo> typeinfo_table{
-  {"uint32", TypeInfo::kUInt32},
-  {"float32", TypeInfo::kFloat32},
-  {"float64", TypeInfo::kFloat64}
 };
 
 }  // namespace treelite
